Embed Storybook's Webpack configuration in its own
