+++ packages/fuselage/package.json
@@ -38,8 +38,8 @@
     "update-storybook": "cross-env NODE_ENV=loki run-s build-storybook loki:update"
   },
   "peerDependencies": {
-    "@rocket.chat/fuselage-hooks": "^0.13.2",
-    "@rocket.chat/icons": "^0.13.2",
+    "@rocket.chat/fuselage-hooks": "*",
+    "@rocket.chat/icons": "*",
     "react": "^16.12.0",
     "react-dom": "^16.12.0"
   },
